Adding support for Options
